requirements.txt
fastapi
uvicorn
python-dotenv
beautifulsoup4
celery
elasticsearch